reorganise docs and examples
