+++ docs/cli/command-line-ref.md
@@ -1,0 +1,25 @@
+segysak - CLI
+===============
+
+segysak offers an interface on the command line to inspect SEG-Y files and to
+convert between data formats.
+
+**segysak**
+
+.. program-output:: segysak
+
+**segysak ebcidc**
+
+.. program-output:: segysak ebcidc --help
+
+**segysak scan**
+
+.. program-output:: segysak scan --help
+
+**segysak scrape**
+
+.. program-output:: segysak scrape --help
+
+**segysak convert**
+
+.. program-output:: segysak convert --help
